requirements.txt
Django==2.1.5
django-bootstrap-breadcrumbs==0.9.1
django-redis==4.10.0
mysqlclient==1.3.13
pytz==2018.7
redis==3.0.1
six==1.12.0
six==1.12.0
django-constance=2.3.1